Add comment autorefresh feature to APIView
