+++ src/dotnet/APIView/APIViewWeb/Controllers/CommentsController.cs
@@ -1,23 +1,30 @@
 using System;
 using System.Threading.Tasks;
+using APIViewWeb.DTO;
 using APIViewWeb.Hubs;
 using APIViewWeb.Managers;
 using APIViewWeb.Models;
+using Azure.Identity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Razor.TagHelpers;
 using Microsoft.AspNetCore.SignalR;
+using Microsoft.Azure.Cosmos.Linq;
+using Microsoft.Extensions.Azure;
+using Octokit;
 
 namespace APIViewWeb.Controllers
 {
-    [Authorize("RequireOrganization")]
+    //[Authorize("RequireOrganization")]
+    [AllowAnonymous]
     public class CommentsController: Controller
     {
         private readonly ICommentsManager _commentsManager;
         private readonly IReviewManager _reviewManager;
-        private readonly IHubContext<NotificationHub> _notificationHubContext;
+        private readonly IHubContext<SignalRHub> _notificationHubContext;
         private readonly INotificationManager _notificationManager;
 
-        public CommentsController(ICommentsManager commentsManager, IReviewManager reviewManager, INotificationManager notificationManager, IHubContext<NotificationHub> notificationHub)
+        public CommentsController(ICommentsManager commentsManager, IReviewManager reviewManager, INotificationManager notificationManager, IHubContext<SignalRHub> notificationHub)
         {
             _notificationHubContext = notificationHub;
             _commentsManager = commentsManager;
@@ -26,7 +33,7 @@ namespace APIViewWeb.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> Add(string reviewId, string revisionId, string elementId, string commentText, string sectionClass, string groupNo, string[] taggedUsers, string resolutionLock = "off", bool usageSampleComment = false)
+        public async Task<ActionResult> Add(string reviewId, string revisionId, string elementId, string commentText, string sectionClass, string groupNo, string[] taggedUsers, string resolutionLock = "off", bool usageSampleComment = false, string signalRConnectionId = null)
         {
             if (string.IsNullOrEmpty(commentText))
             {
@@ -58,7 +65,8 @@ namespace APIViewWeb.Controllers
             {
                 await _notificationManager.SubscribeAsync(review,User);
             }
-            return await CommentPartialAsync(reviewId, comment.ElementId);
+            
+            return await CommentPartialAsync(reviewId, comment.ElementId); 
         }
 
         [HttpPost]
