+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Index.cshtml.cs
@@ -20,13 +20,13 @@ namespace APIViewWeb.Pages.Assemblies
     public class IndexPageModel : PageModel
     {
         private readonly IReviewManager _manager;
-        private readonly IHubContext<NotificationHub> _notificationHubContext;
+        private readonly IHubContext<SignalRHub> _notificationHubContext;
         public readonly UserPreferenceCache _preferenceCache;
         public readonly IUserProfileManager _userProfileManager;
         public const int _defaultPageSize = 50;
         public const string _defaultSortField = "LastUpdated";
 
-        public IndexPageModel(IReviewManager manager, IUserProfileManager userProfileManager, UserPreferenceCache preferenceCache, IHubContext<NotificationHub> notificationHub)
+        public IndexPageModel(IReviewManager manager, IUserProfileManager userProfileManager, UserPreferenceCache preferenceCache, IHubContext<SignalRHub> notificationHub)
         {
             _notificationHubContext = notificationHub;
             _manager = manager;
